requirements.txt
Django>=4.0.4
gunicorn>=20.1.0
redis>=4.3.4
celery>=5.2.7
yt-dlp==2023.10.7
django-celery-beat>=2.3.0
django-cleanup>=6.0.0
psycopg2-binary>=2.9.3
gevent>=23.9.1
python-dotenv